requirements.txt
configparser==3.5.0
Django==1.11.6
django-filter==1.0.4
djangorestframework==3.7.1
Markdown==2.6.9
mysqlclient==1.3.12
pkg-resources==0.0.0
pytz==2017.2
dj-database-url==0.4.2
dj-static==0.0.6
